% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoshiny.R
\name{neoshiny}
\alias{neoshiny}
\title{Starts shiny application for the neodistr package}
\usage{
neoshiny()
}
\value{
Starts shiny application for the neodistr package.
}
\description{
Starts shiny application for the neodistr package
}
\examples{
if 	(interactive()) {
	suppressMessages(library(neodistr))	
	neoshiny()
	}
}
\author{
Anisa' Faoziah and Achmad Syahrul Choir
}
